+++ lib/main.dart
@@ -14,7 +14,7 @@ import 'core/logger.dart';
 import 'core/progress_manager.dart';
 import 'utils/web_utils.dart';
 import 'utils/wechat_adapter.dart';
-import 'utils/web_storage.dart';
+
 import 'utils/performance_optimizer.dart';
 import 'utils/storage_adapter.dart';
 import 'widgets/header.dart';